+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog] and Pipelines adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [unreleased]
+
 ## [0.0.65] - 2022-04-24
 ### Add
 - `--show` and `--show-pipelines`: Annotate steps' conditions with `*C`