+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog] and Pipelines adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [unreleased]
+## [0.0.66] - 2022-06-22
 ### Add
 - Run step scripts with `/bin/bash` if it is available; disable with
   `script.bash-runner=false` for previous behaviour (#17)  (thanks
@@ -696,4 +696,5 @@ The format is based on [Keep a Changelog] and Pipelines adheres to
 [0.0.63]: https://github.com/ktomk/pipelines/releases/tag/0.0.63
 [0.0.64]: https://github.com/ktomk/pipelines/releases/tag/0.0.64
 [0.0.65]: https://github.com/ktomk/pipelines/releases/tag/0.0.65
+[0.0.66]: https://github.com/ktomk/pipelines/releases/tag/0.0.66
 [unreleased]: https://github.com/ktomk/pipelines